+++ src/categories/categories.gateway.ts
@@ -10,16 +10,16 @@ import {
     WebSocketServer,
 } from '@nestjs/websockets'
 import { Server, Socket } from 'socket.io'
+import { SocketIoCurrentUser } from 'src/auth/decorators/socket.io-current-user.decorator'
 import { SocketIoJwtAuthGuard } from 'src/auth/guards/socket.io-jwt.guard'
+import { UserFromRequest } from 'src/auth/types/request-response'
 import { TokenPayload } from 'src/auth/types/tokenPayload'
-import { InternalCategoriesCreatedEvent } from 'src/categories/events/internal-categories-created.event'
-import { InternalCategoriesUpdatedEvent } from 'src/categories/events/internal-categories-updated.event'
+import { InternalCategoriesCudEvent } from 'src/categories/events/internal-categories.CUD.event'
 import { SocketIoService } from 'src/socketio/socketio.service'
 import { User } from 'src/users/models/users.model'
 import { UsersService } from 'src/users/users.service'
 import { CategoriesService } from './categories.service'
 import { CreateCategoryDto } from './dto/create-category.dto'
-import { InternalCategoriesDeletedEvent } from './events/internal-categories-deleted.event'
 import { Category } from './models/categories.model'
 
 @WebSocketGateway(8080, { cors: { origin: '*' }, namespace: '/categories' })
@@ -67,10 +67,11 @@ export class CategoriesGateway implements OnGatewayConnection, OnGatewayDisconne
     @UseGuards(SocketIoJwtAuthGuard)
     async updateCategory(
         @ConnectedSocket() socket: Socket,
-        @MessageBody() { name, categoryId }: { name: string, categoryId: string}
+        @MessageBody() { name, categoryId }: { name: string, categoryId: string},
+        @SocketIoCurrentUser() user: UserFromRequest,
     ): Promise<void> {
         const category: Category = await this.categoriesService.getCategoryById(categoryId)
-        await this.categoriesService.updateCategory({ name, category })
+        await this.categoriesService.updateCategory({ name, category, userId: user.id })
         socket.emit('200', category)
     }
 
@@ -78,15 +79,16 @@ export class CategoriesGateway implements OnGatewayConnection, OnGatewayDisconne
     @UseGuards(SocketIoJwtAuthGuard)
     async deleteCategory(
         @ConnectedSocket() socket: Socket,
-        @MessageBody() { categoryId }: { categoryId: string}
+        @MessageBody() { categoryId }: { categoryId: string},
+        @SocketIoCurrentUser() user: UserFromRequest,
     ): Promise<void> {
         const category: Category = await this.categoriesService.getCategoryById(categoryId)
-        await this.categoriesService.deleteCategory({ category })
+        await this.categoriesService.deleteCategory({ category, userId: user.id })
         socket.emit('200', category)
     }
 
     @OnEvent('internal-categories.created')
-    async showToSocketsNewCategory(event: InternalCategoriesCreatedEvent): Promise<void>  {
+    async showToSocketsNewCategory(event: InternalCategoriesCudEvent): Promise<void>  {
         const sockets = await this.server.fetchSockets()
         const socketsOfGroupUsers = (await this.socketIoService.getClients())
             .filter(client => event.usersIds.some(userId => userId === client.userId))
@@ -97,7 +99,7 @@ export class CategoriesGateway implements OnGatewayConnection, OnGatewayDisconne
     }
 
     @OnEvent('internal-categories.updated')
-    async showToSocketsUpdatedCategory(event: InternalCategoriesUpdatedEvent): Promise<void>  {
+    async showToSocketsUpdatedCategory(event: InternalCategoriesCudEvent): Promise<void>  {
         const socketsOfGroupUsers = (await this.socketIoService.getSocketsByUsersIds(
             (await this.server.fetchSockets()),
             event.usersIds
@@ -108,14 +110,14 @@ export class CategoriesGateway implements OnGatewayConnection, OnGatewayDisconne
     }
 
     @OnEvent('internal-categories.deleted')
-    async hideFromSocketsDeletedCategory(event: InternalCategoriesDeletedEvent): Promise<void>  {
+    async hideFromSocketsDeletedCategory(event: InternalCategoriesCudEvent): Promise<void>  {
         const sockets = await this.server.fetchSockets()
         const socketsOfGroupUsers = (await this.socketIoService.getClients())
             .filter(client => event.usersIds.some(userId => userId === client.userId))
             .map(client => sockets.find(socket => socket.id === client.socketId))
         this.server
             .to(socketsOfGroupUsers.map(socket => socket.id))
-            .emit('category-deleted', event.categoryId)
+            .emit('category-deleted', event.category)
     }
 
 }
